Higher precision for some filters, add bitwise ops for bool sigs

+++ modules/src/low_level/biquad_band_pass_filter.rs
@@ -9,16 +9,16 @@
 
 #[derive(Default, Debug)]
 struct BufferEntry {
-    a: f32,
-    d1: f32,
-    d2: f32,
-    d3: f32,
-    d4: f32,
-    w0: f32,
-    w1: f32,
-    w2: f32,
-    w3: f32,
-    w4: f32,
+    a: f64,
+    d1: f64,
+    d2: f64,
+    d3: f64,
+    d4: f64,
+    w0: f64,
+    w1: f64,
+    w2: f64,
+    w3: f64,
+    w4: f64,
 }
 
 #[derive(Debug)]
@@ -35,7 +35,7 @@ impl Buffer {
         Self { entries }
     }
 
-    fn apply(&mut self, mut sample: f32) -> f32 {
+    fn apply(&mut self, mut sample: f64) -> f64 {
         for entry in self.entries.iter_mut() {
             entry.w0 = (entry.d1 * entry.w1)
                 + (entry.d2 * entry.w2)
@@ -60,12 +60,12 @@ impl Buffer {
 
 pub mod butterworth {
     use super::*;
-    use std::f32::consts::PI;
+    use std::f64::consts::PI;
 
     pub struct State {
         buffer: Buffer,
-        prev_lower_half_power_frequency_hz: f32,
-        prev_upper_half_power_frequency_hz: f32,
+        prev_lower_half_power_frequency_hz: f64,
+        prev_upper_half_power_frequency_hz: f64,
     }
 
     impl State {
@@ -85,11 +85,11 @@ pub mod butterworth {
 
         pub fn run(
             &mut self,
-            sample: f32,
-            sample_rate_hz: f32,
-            lower_half_power_frequency_hz: f32,
-            upper_half_power_frequency_hz: f32,
-        ) -> f32 {
+            sample: f64,
+            sample_rate_hz: f64,
+            lower_half_power_frequency_hz: f64,
+            upper_half_power_frequency_hz: f64,
+        ) -> f64 {
             let lower_half_power_frequency_hz =
                 lower_half_power_frequency_hz.max(0.0);
             let upper_half_power_frequency_hz = upper_half_power_frequency_hz
@@ -127,9 +127,9 @@ pub mod butterworth {
 
     fn update_entries(
         buffer: &mut Buffer,
-        lower_half_power_frequency_hz: f32,
-        upper_half_power_frequency_hz: f32,
-        sample_rate_hz: f32,
+        lower_half_power_frequency_hz: f64,
+        upper_half_power_frequency_hz: f64,
+        sample_rate_hz: f64,
     ) {
         let a = ((PI
             * (lower_half_power_frequency_hz + upper_half_power_frequency_hz))
@@ -146,9 +146,9 @@ pub mod butterworth {
             / sample_rate_hz)
             .tan();
         let b2 = b * b;
-        let n = buffer.entries.len() as f32;
+        let n = buffer.entries.len() as f64;
         for (i, entry) in buffer.entries.iter_mut().enumerate() {
-            let r = ((PI * ((2.0 * i as f32) + 1.0)) / (4.0 * n)).sin();
+            let r = ((PI * ((2.0 * i as f64) + 1.0)) / (4.0 * n)).sin();
             let s = b2 + (2.0 * b * r) + 1.0;
             entry.a = b2 / s;
             entry.d1 = 4.0 * a * (1.0 + (b * r)) / s;
@@ -161,15 +161,15 @@ pub mod butterworth {
 
 pub mod chebyshev {
     use super::*;
-    use std::f32::consts::PI;
+    use std::f64::consts::PI;
 
-    pub const EPSILON_MIN: f32 = 0.01;
+    pub const EPSILON_MIN: f64 = 0.01;
 
     pub struct State {
         buffer: Buffer,
-        prev_lower_cutoff_hz: f32,
-        prev_upper_cutoff_hz: f32,
-        prev_epsilon: f32,
+        prev_lower_cutoff_hz: f64,
+        prev_upper_cutoff_hz: f64,
+        prev_epsilon: f64,
     }
 
     impl State {
@@ -190,12 +190,12 @@ pub mod chebyshev {
 
         pub fn run(
             &mut self,
-            sample: f32,
-            sample_rate_hz: f32,
-            lower_cutoff_hz: f32,
-            upper_cutoff_hz: f32,
-            epsilon: f32,
-        ) -> f32 {
+            sample: f64,
+            sample_rate_hz: f64,
+            lower_cutoff_hz: f64,
+            upper_cutoff_hz: f64,
+            epsilon: f64,
+        ) -> f64 {
             if self.buffer.entries.is_empty() {
                 return sample;
             }
@@ -231,10 +231,10 @@ pub mod chebyshev {
 
     fn update_entries(
         buffer: &mut Buffer,
-        lower_cutoff_hz: f32,
-        upper_cutoff_hz: f32,
-        epsilon: f32,
-        sample_rate_hz: f32,
+        lower_cutoff_hz: f64,
+        upper_cutoff_hz: f64,
+        epsilon: f64,
+        sample_rate_hz: f64,
     ) {
         let a = ((PI * (lower_cutoff_hz + upper_cutoff_hz)) / sample_rate_hz)
             .cos()
@@ -245,11 +245,11 @@ pub mod chebyshev {
             ((PI * (upper_cutoff_hz - lower_cutoff_hz)) / sample_rate_hz).tan();
         let b2 = b * b;
         let u = ((1.0 + (1.0 + (epsilon * epsilon)).sqrt()) / epsilon).ln();
-        let n = (buffer.entries.len() * 4) as f32;
+        let n = (buffer.entries.len() * 4) as f64;
         let su = ((2.0 * u) / n).sinh();
         let cu = ((2.0 * u) / n).cosh();
         for (i, entry) in buffer.entries.iter_mut().enumerate() {
-            let theta = (PI * ((2.0 * i as f32) + 1.0)) / n;
+            let theta = (PI * ((2.0 * i as f64) + 1.0)) / n;
             let r = theta.sin() * su;
             let c = theta.cos() * cu;
             let c = (r * r) + (c * c);
