+++ modules/src/reverb_freeverb.rs
@@ -41,11 +41,11 @@ builder! {
     pub struct PropsBuilder {
         #[generic_with_constraint = "SigT<Item = f32>"]
         #[generic_name = "R"]
-        #[default = freeverb::INITIAL_ROOM_SIZE]
+        #[default = freeverb::INITIAL_ROOM_SIZE as f32]
         room_size: f32,
         #[generic_with_constraint = "SigT<Item = f32>"]
         #[generic_name = "D"]
-        #[default = freeverb::INITIAL_DAMPING]
+        #[default = freeverb::INITIAL_DAMPING as f32]
         damping: f32,
         #[generic_with_constraint = "SigT<Item = f32>"]
         #[generic_name = "M"]
@@ -90,8 +90,8 @@ where
             props,
             sig,
             state: freeverb::ReverbModel::new(),
-            room_size_prev: freeverb::INITIAL_ROOM_SIZE,
-            damping_prev: freeverb::INITIAL_DAMPING,
+            room_size_prev: freeverb::INITIAL_ROOM_SIZE as f32,
+            damping_prev: freeverb::INITIAL_DAMPING as f32,
             buf: Vec::new(),
         }
     }
@@ -117,13 +117,13 @@ where
         } {
             if room_size != self.room_size_prev {
                 self.room_size_prev = room_size;
-                self.state.set_room_size(room_size);
+                self.state.set_room_size(room_size as f64);
             }
             if damping != self.damping_prev {
                 self.damping_prev = damping;
-                self.state.set_damping(damping);
+                self.state.set_damping(damping as f64);
             }
-            let reverb = self.state.process(sample);
+            let reverb = self.state.process(sample as f64) as f32;
             *out = (reverb * mix) + (sample * (1.0 - mix));
         }
         &self.buf
